Refactored code to separate endpoint implementations
