+++ app/apis/auth/services/update_profile_service.py
@@ -1,0 +1,37 @@
+from typing import Union
+
+from apis.auth.utils import get_current_user, get_user_by_username
+from db.models import User
+from db.session import get_db
+from fastapi import APIRouter, Depends, status
+from pydantic import BaseModel, Extra
+from sqlalchemy.orm import Session
+from typing_extensions import Annotated
+
+router = APIRouter()
+
+
+class UserUpdate(BaseModel):
+    username: str
+    first_name: Union[str, None] = None
+    last_name: Union[str, None] = None
+    phone_number: Union[str, None] = None
+
+
+@router.put("/profile", response_model=UserUpdate, status_code=status.HTTP_200_OK)
+def update_profile(
+    user: UserUpdate,
+    current_user: Annotated[User, Depends(get_current_user)],
+    db: Session = Depends(get_db),
+):
+    db_user = get_user_by_username(db, user.username)
+
+    for var, value in user.dict().items():
+        if value:
+            setattr(db_user, var, value)
+
+    db.add(db_user)
+    db.commit()
+    db.refresh(db_user)
+
+    return db_user
