+++ app/apis/menu/services/delete_menu_item_service.py
@@ -1,0 +1,21 @@
+from typing import List
+
+from apis.auth.utils import RolesBasedAuthChecker, get_current_user
+from apis.menu import utils
+from db.models import User, UserRole
+from db.session import get_db
+from fastapi import APIRouter, Depends, status
+from sqlalchemy.orm import Session
+from typing_extensions import Annotated
+
+router = APIRouter()
+
+
+@router.delete("/menu/{item_id}", status_code=status.HTTP_204_NO_CONTENT)
+def delete_menu_item(
+    item_id: int,
+    current_user: Annotated[User, Depends(get_current_user)],
+    db: Session = Depends(get_db),
+    # auth=Depends(RolesBasedAuthChecker([UserRole.EMPLOYEE, UserRole.CHEF])),
+):
+    utils.delete_menu_item(db, item_id)
