+++ app/apis/menu/services/get_menu_service.py
@@ -1,0 +1,16 @@
+from typing import List
+
+from apis.auth.utils import RolesBasedAuthChecker, get_current_user
+from apis.menu import schemas, utils
+from db.models import MenuItem, User, UserRole
+from db.session import get_db
+from fastapi import APIRouter, Depends, status
+from sqlalchemy.orm import Session
+from typing_extensions import Annotated
+
+router = APIRouter()
+
+
+@router.get("/menu", response_model=List[schemas.MenuItem])
+def get_menu(db: Session = Depends(get_db)):
+    return db.query(MenuItem).all()
