+++ app/apis/orders/services/get_order_service.py
@@ -1,0 +1,20 @@
+from apis.auth.utils import RolesBasedAuthChecker
+from apis.orders import schemas
+from db.models import Order, UserRole
+from db.session import get_db
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy.orm import Session
+
+router = APIRouter()
+
+
+@router.get("/orders/{order_id}", response_model=schemas.Order)
+def get_order(
+    order_id: int,
+    db: Session = Depends(get_db),
+    auth=Depends(RolesBasedAuthChecker([UserRole.CUSTOMER])),
+):
+    db_order = db.query(Order).filter(Order.id == order_id).first()
+    if db_order is None:
+        raise HTTPException(status_code=404, detail="Order not found")
+    return db_order
