+++ app/apis/users/services/update_user_role_service.py
@@ -1,0 +1,27 @@
+from apis.auth.utils import get_current_user, update_user
+from apis.users.schemas import UserRoleUpdate
+from db import models
+from db.session import get_db
+from fastapi import APIRouter, Depends, HTTPException, status
+from sqlalchemy.orm import Session
+from typing_extensions import Annotated
+
+router = APIRouter()
+
+
+@router.put("/users/update_role", response_model=UserRoleUpdate)
+async def update_user_role(
+    user: UserRoleUpdate,
+    current_user: Annotated[models.User, Depends(get_current_user)],
+    db: Session = Depends(get_db),
+):
+    # this method allows staff to give Employee role to other users
+    # Chef role is restricted
+    if user.role == models.UserRole.CHEF.value:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Only Chef is authorized to add Chef role!",
+        )
+
+    db_user = update_user(db, user.username, user)
+    return current_user
