+++ app/tests/modules/menu/test_menu_service.py
@@ -49,7 +49,7 @@ def test_create_menu_item_by_employee_or_chef_returns_201(
     assert response.status_code == 201
 
 
-def test_delete_menu_item_by_employee_or_chef_returns_204(
+def test_delete_menu_item_by_employee_or_chef_should_return_204(
     test_db, employee_client, chef_client
 ):
     menu_item = MenuItem(
@@ -69,27 +69,9 @@ def test_delete_menu_item_by_employee_or_chef_returns_204(
     assert response.status_code == 204
 
 
-def test_delete_menu_item_by_employee_or_chef_returns_204(
-    test_db, employee_client, chef_client
+def test_delete_menu_item_by_unauthenticated_user_should_return_401(
+    test_db, anon_client
 ):
-    menu_item = MenuItem(
-        name="Item", price=10.99, category="", description="", image_base64=""
-    )
-    test_db.add(menu_item)
-    test_db.commit()
-    response = employee_client.delete(f"/menu/{menu_item.id}")
-    assert response.status_code == 204
-
-    menu_item = MenuItem(
-        name="Item", price=10.99, category="", description="", image_base64=""
-    )
-    test_db.add(menu_item)
-    test_db.commit()
-    response = chef_client.delete(f"/menu/{menu_item.id}")
-    assert response.status_code == 204
-
-
-def test_delete_menu_item_by_unauthorized_user_returns_401(test_db, anon_client):
     menu_item = MenuItem(
         name="Item", price=10.99, category="", description="", image_base64=""
     )
